content update with message and crossorigin

+++ 14_Spring_Data_JPA/src/main/java/lk/ijse/spring/service/impl/CustomerServiceImpl.java
@@ -4,6 +4,7 @@ import lk.ijse.spring.dto.CustomerDTO;
 import lk.ijse.spring.entity.Customer;
 import lk.ijse.spring.repo.CustomerRepo;
 import lk.ijse.spring.service.CustomerService;
+import net.bytebuddy.description.method.MethodDescription;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional// AOP
@@ -22,12 +24,14 @@ public class CustomerServiceImpl implements CustomerService {
     @Autowired
     ModelMapper mapper;
 
-    public CustomerServiceImpl() {
-        System.out.println("CustomerServiceImpl Instantiated");
-    }
 
     @Override
     public void addCustomer(CustomerDTO dto) {
+        //service level validations
+        if (customerRepo.existsById(dto.getId())) {
+            throw new RuntimeException(dto.getId()+" is already available, please insert a new ID");
+        }
+
         Customer map = mapper.map(dto, Customer.class);
         //first param = source
         //Type you want to convert
@@ -36,6 +40,9 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomer(String id) {
+        if (!customerRepo.existsById(id)) {
+            throw new RuntimeException(id+ " Customer is not available, please check the ID before delete.!");
+        }
         customerRepo.deleteById(id);
     }
 
@@ -50,12 +57,18 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public CustomerDTO findCustomer(String id) {
+        if (!customerRepo.existsById(id)) {
+            throw new RuntimeException(id+ " Customer is not available, please check the ID.!");
+        }
         Customer customer = customerRepo.findById(id).get();
         return mapper.map(customer,CustomerDTO.class);
     }
 
     @Override
     public void updateCustomer(CustomerDTO c) {
+        if (!customerRepo.existsById(c.getId())) {
+            throw new RuntimeException(c.getId()+ " Customer is not available, please check the ID before update.!");
+        }
         Customer map = mapper.map(c, Customer.class);
         customerRepo.save(map);
     }
